*04* สร้างหลอดเลือดยาน  80%

+++ f2/spw/GameReporter.java
@@ -4,6 +4,6 @@ public interface GameReporter {
 
 	long getScore();
 	
-	long getHp();
+	int getHp();
 
 }
